[refactor] 'hide' some JavaSupport internals
to avoid leaking it thought JRuby's code-base
