+++ core/src/main/java/org/jruby/javasupport/JavaSupport.java
@@ -34,19 +34,60 @@
 
 package org.jruby.javasupport;
 
+import org.jruby.Ruby;
 import org.jruby.RubyClass;
 import org.jruby.RubyModule;
 import org.jruby.javasupport.binding.AssignedName;
+import org.jruby.javasupport.ext.JavaExtensions;
 import org.jruby.javasupport.proxy.JavaProxyClass;
 import org.jruby.javasupport.util.ObjectProxyCache;
 import org.jruby.runtime.builtin.IRubyObject;
 import org.jruby.util.collections.ClassValue;
+import org.jruby.util.collections.ClassValueCalculator;
 
 import java.lang.reflect.Member;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.locks.ReentrantLock;
 
 public abstract class JavaSupport {
+
+    protected final Ruby runtime;
+
+    private final ClassValue<JavaClass> javaClassCache;
+    private final ClassValue<RubyModule> proxyClassCache;
+
+    static final class UnfinishedProxy extends ReentrantLock {
+        final RubyModule proxy;
+        UnfinishedProxy(RubyModule proxy) {
+            this.proxy = proxy;
+        }
+    }
+
+    private final Map<Class, UnfinishedProxy> unfinishedProxies;
+
+    protected JavaSupport(final Ruby runtime) {
+        this.runtime = runtime;
+
+        this.javaClassCache = ClassValue.newInstance(klass -> new JavaClass(runtime, getJavaClassClass(), klass));
+
+        this.proxyClassCache = ClassValue.newInstance(new ClassValueCalculator<RubyModule>() {
+            /**
+             * Because of the complexity of processing a given class and all its dependencies,
+             * we opt to synchronize this logic. Creation of all proxies goes through here,
+             * allowing us to skip some threading work downstream.
+             */
+            @Override
+            public synchronized RubyModule computeValue(Class<?> klass) {
+                RubyModule proxyKlass = Java.createProxyClassForClass(runtime, klass);
+                JavaExtensions.define(runtime, klass, proxyKlass); // (lazy) load extensions
+                return proxyKlass;
+            }
+        });
+        // Proxy creation is synchronized (see above) so a HashMap is fine for recursion detection.
+        this.unfinishedProxies = new ConcurrentHashMap<>(8, 0.75f, 1);
+    }
 
     public Class loadJavaClass(String className) throws ClassNotFoundException {
         return loadJavaClass(className, true);
@@ -56,10 +97,6 @@ public abstract class JavaSupport {
     public abstract Class loadJavaClassVerbose(String className);
 
     public abstract Class loadJavaClassQuiet(String className);
-
-    public abstract JavaClass getJavaClassFromCache(Class clazz);
-
-    public abstract RubyModule getProxyClassFromCache(Class clazz);
 
     public abstract void handleNativeException(Throwable exception, Member target);
 
@@ -191,18 +228,34 @@ public abstract class JavaSupport {
         }
     }
 
-    /**
-     * @deprecated Internal API that should not be accessible.
-     */
-    public abstract void beginProxy(Class cls, RubyModule proxy);
+    // Internal API
+
+    final void beginProxy(Class clazz, RubyModule proxy) {
+        UnfinishedProxy up = new UnfinishedProxy(proxy);
+        up.lock();
+        unfinishedProxies.put(clazz, up);
+    }
+
+    final void endProxy(Class clazz) {
+        UnfinishedProxy up = unfinishedProxies.remove(clazz);
+        up.unlock();
+    }
+
+    final RubyModule getUnfinishedProxy(Class clazz) {
+        UnfinishedProxy up = unfinishedProxies.get(clazz);
+        if (up != null && up.isHeldByCurrentThread()) return up.proxy;
+        return null;
+    }
+
+    RubyModule getProxyClassFromCache(Class clazz) {
+        return proxyClassCache.get(clazz);
+    }
 
     /**
-     * @deprecated Internal API that should not be accessible.
+     * @deprecated Internal API - no longer used
      */
-    public abstract void endProxy(Class cls);
+    public JavaClass getJavaClassFromCache(Class clazz) {
+        return javaClassCache.get(clazz);
+    }
 
-    /**
-     * @deprecated Internal API that should not be accessible.
-     */
-    public abstract RubyModule getUnfinishedProxy(Class cls);
 }
